requirements.txt
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.8.0
Flask==3.0.0
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
frozendict==2.3.10
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.2
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==6.11.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multitasking==0.0.11
mysql-connector-python==8.2.0
mysqlclient==2.2.1
numpy==1.26.2
openai==1.5.0
packaging==23.2
pandas==2.1.4
pdfkit==1.0.0
peewee==3.17.0
Pillow==10.1.0
protobuf==4.21.12
pyarrow==14.0.1
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
Pygments==2.17.2
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.14.1
setuptools==69.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
streamlit==1.29.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.25.0
validators==0.22.0
waitress==2.1.2
watchdog==3.0.0
webencodings==0.5.1
Werkzeug==3.0.1
wheel==0.42.0
yfinance==0.2.33
zipp==3.17.0